backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
httpx
qrcode[pil]
python-dotenv
pydantic